[task] put the types back in identity
They had been moved to a separate types file. Now they're back.
